refactor: replace bare-ancestor barrel imports
