+++ packages/salesforcedx-utils-vscode/src/helpers/paths.ts
@@ -8,7 +8,7 @@
 import { Global } from '@salesforce/core/global';
 import * as path from 'node:path';
 import { URI } from 'vscode-uri';
-import { WorkspaceContextUtil } from '..';
+import { WorkspaceContextUtil } from '../context/workspaceContextUtil';
 import { workspaceUtils } from '../workspaces/workspaceUtils';
 import { createDirectory } from './fs';
