+++ packages/salesforcedx-utils-vscode/src/telemetry/reporters/telemetryFile.ts
@@ -8,9 +8,9 @@ import type { TelemetryReporter } from '@salesforce/vscode-service-provider';
 import * as path from 'node:path';
 import * as vscode from 'vscode';
 import { URI } from 'vscode-uri';
-import { getRootWorkspacePath } from '../..';
 import { LOCAL_TELEMETRY_FILE } from '../../constants';
 import { WorkspaceContextUtil } from '../../context/workspaceContextUtil';
+import { getRootWorkspacePath } from '../../workspaces/workspaceUtils';
 
 /**
  * Represents a telemetry file that logs telemetry events by appending to a local file.
