+++ packages/salesforcedx-visualforce-markup-language-server/src/services/htmlLinks.ts
@@ -6,8 +6,11 @@
 
 import { Range, TextDocument, DocumentLink } from 'vscode-languageserver-types';
 import { URI } from 'vscode-uri';
-import { DocumentContext } from '..';
 import { createScanner, TokenType } from '../parser/htmlScanner';
+
+type DocumentContext = {
+  resolveReference(ref: string, base?: string): string;
+};
 
 function stripQuotes(url: string): string {
   return url.replace(/^'([^']*)'$/, (substr, match1) => match1).replace(/^"([^"]*)"$/, (substr, match1) => match1);
