+++ packages/salesforcedx-vscode-core/src/context/workspaceContext.ts
@@ -17,7 +17,7 @@ import {
 import * as Effect from 'effect/Effect';
 import * as vscode from 'vscode';
 import { getRuntime } from '../services/runtime';
-import { workspaceContextUtils } from '.';
+import { getOrgShape } from './workspaceOrgShape';
 
 /**
  * Manages the context of a workspace during a session with an open SFDX Project.
@@ -69,7 +69,7 @@ export class WorkspaceContext {
   protected async handleOrgShapeChange(orgInfo: OrgUserInfo) {
     const { username } = orgInfo;
     if (username !== undefined) {
-      const orgShape = await workspaceContextUtils.getOrgShape(username);
+      const orgShape = await getOrgShape(username);
       if (orgShape !== 'Undefined') {
         WorkspaceContextUtil.getInstance().orgShape = orgShape;
         WorkspaceContextUtil.getInstance().devHubId = undefined;
